% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{AUC_synergy}
\alias{AUC_synergy}
\title{Calculatr AUC based synergy scores and its bootstrap confidence interval}
\usage{
AUC_synergy(
  auc_lst,
  t = 21,
  method = "Bliss",
  boot_n = 1000,
  ci = 0.95,
  ci_type = "perc",
  kw = "Test",
  file = "eGR_synergy",
  save = T,
  display = F,
  parallel = "no"
)
}
\arguments{
\item{auc_lst}{results from AUC calculation}

\item{t}{date to estimate survival (default 21 days)}

\item{method}{method for synergy calculation, can be Bliss,HSA or RA}

\item{boot_n}{number of bootstrap resample}

\item{ci}{confidence intervaln for AUC based synergy score, can be 0.95,0.9,0.8, etc}

\item{ci_type}{type of bootstrap confidence interval, can be stud,perc or bca}

\item{kw}{legend keyword for plotting}

\item{file}{if save is TRUE, the name of output file}

\item{save}{save image}

\item{display}{print image}

\item{parallel}{whether or not usint parallel computing}
}
\value{
a dataframe of synergy scores and its bootstrap confidence interval
}
\description{
Calculatr AUC based synergy scores and its bootstrap confidence interval
}
\examples{
auc_lst=get_mAUCr(LS_1034)
bliss_synergy_AUC=AUC_synergy(auc_lst)
}
