% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy.R
\name{get_AUC}
\alias{get_AUC}
\title{Calculate normalized AUC for tumor growth curve}
\usage{
get_AUC(df)
}
\arguments{
\item{df}{tumor growth data for a single mouse}
}
\value{
an AUC dataframe
}
\description{
Calculate normalized AUC for tumor growth curve
}
\examples{
get_AUC(mouse_tv)
}
