% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCI}
\alias{getCI}
\title{Get bootstrap confidence interval}
\usage{
getCI(res, i, conf = 0.95, ci_type = "perc")
}
\arguments{
\item{res}{bootstrap results}

\item{i}{index of bootstrap statistic}

\item{conf}{confidence interval}

\item{ci_type}{type of bootstrap confidence interval, can be perc,stud or bca}
}
\value{
boostrap confidence interval
}
\description{
Get bootstrap confidence interval
}
