% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{global_CI_synergy}
\alias{global_CI_synergy}
\title{Calculatr global CI from CombPDX and its bootstrap confidence interval}
\usage{
global_CI_synergy(
  tv,
  method = "Bliss",
  boot_n = 1000,
  ci = 0.95,
  ci_type = "perc",
  display = TRUE,
  save = TRUE
)
}
\arguments{
\item{tv}{tumor growth data}

\item{method}{method for synergy calculation, can be Bliss,HSA or RA}

\item{boot_n}{number of bootstrap resample}

\item{ci}{confidence interval for CombPDX global CI, can be 0.95,0.9,0.8, etc}

\item{ci_type}{type of bootstrap confidence interval, can be stud,perc or bca}

\item{display}{print image}

\item{save}{save image}
}
\value{
a dataframe of synergy scores and its bootstrap confidence interval
}
\description{
Calculatr global CI from CombPDX and its bootstrap confidence interval
}
