% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy.R
\name{mAUCr}
\alias{mAUCr}
\title{Calculate medianAUCRatio for two groups}
\usage{
mAUCr(df, grp, ref_grp)
}
\arguments{
\item{df}{AUC data frame}

\item{grp}{grp id}

\item{ref_grp}{reference group id}
}
\value{
A number
}
\description{
Calculate medianAUCRatio for two groups
}
