% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_density}
\alias{plot_density}
\title{Density plot for selected variable in a data frame}
\usage{
plot_density(df, sel_var, cutoff = NA, pval = NA, pe = NA, lb = NA, ub = NA)
}
\arguments{
\item{df}{data frame}

\item{sel_var}{variable to plot}

\item{pval}{p value (optional)}

\item{pe}{point estimate}

\item{lb}{lower confidence interval}

\item{ub}{higher confidence interval}
}
\value{
a ggplot object
}
\description{
Density plot for selected variable in a data frame
}
