% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_group_by_day}
\alias{plot_group_by_day}
\title{Draw boxplot to compare group means for each day}
\usage{
plot_group_by_day(tv, y = "TV", save = TRUE)
}
\arguments{
\item{tv}{object return from read_tv}

\item{y}{variable to plot,can be TV, RTV, DeltaTV or logTV}

\item{save}{whether or not save the plot}
}
\value{
a ggplot object
}
\description{
Draw boxplot to compare group means for each day
}
\examples{
plot_group_by_day(LS_1034,y='RTV')
}
