% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tumor_growth_curve}
\alias{plot_tumor_growth_curve}
\title{Draw tumor growth curve}
\usage{
plot_tumor_growth_curve(tv, y = "TV", save = TRUE, kw = "Test")
}
\arguments{
\item{tv}{object return from read_tv}

\item{y}{variable to plot,can be TV, RTV, DeltaTV or logTV}

\item{save}{whether or not save the plot}
}
\value{
a ggplot object
}
\description{
Draw tumor growth curve
}
\examples{
plot_tumor_growth_curve(LS_1034)
}
