% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_calc}
\alias{power_calc}
\title{Power calculation for specific design, using simulated data derived from fixed parameters}
\usage{
power_calc(
  method = "Bliss",
  type = "TGI",
  n_sim = 100,
  n = 5,
  tv0 = 50,
  cv_tv0 = 0.2,
  vehicle_double_time = 3,
  t = 21,
  cv_r = 0.05,
  eu_tv = 3000,
  tgi_a = 0.5,
  tgi_b = 0.4,
  tgi_c = 0.8,
  kit_a = 0,
  kit_b = 0,
  kit_c = 0,
  rp_a = 0,
  rp_b = 0,
  rp_c = 0,
  irt_a = 0,
  irt_b = 0,
  irt_c = 0,
  sigma = 1,
  n_core = 1
)
}
\arguments{
\item{method}{Which model to use,can be Bliss, HSA or RA}

\item{type}{Which metric to use, can be AUC or TGI}

\item{n_sim}{number of simulations}

\item{n}{number of animals per group}

\item{tv0}{initial average tumor volume}

\item{cv_tv0}{standard deviation of initial tumor volume, also assuming standard deviation of tumor volume is sd0}

\item{vehicle_double_time}{days to take for double initial tumor size}

\item{t}{length of study}

\item{cv_r}{ratio between standard deviation of tumor growth rate and tumor growth rate for vehicle, reflects heterogeneity of tumor growth}

\item{eu_tv}{mouse is killed if TV exceeds eu_tv}

\item{tgi_a}{TGI of drug a at day 21}

\item{tgi_b}{TGI of durg b at day 21}

\item{tgi_c}{TGI of combo at day 21}

\item{kit_a}{kick in time for drug a, default 0}

\item{kit_b}{kick in time for drug b, default 0}

\item{kit_c}{kick in time for drug combo, default 0}

\item{rp_a}{resistant proportion of drug a}

\item{rp_b}{resistant proportion of drug b}

\item{rp_c}{resistant proportion of drug combo}

\item{irt_a}{day of resistance for drug a (if not 0, induced resistance)}

\item{irt_b}{day of resistance for drug b}

\item{irt_c}{day of resistance for drug c}

\item{sigma}{random error, default to 1}

\item{n_core}{number of cores used for parallel computing}
}
\value{
power of the experiment
}
\description{
Power calculation for specific design, using simulated data derived from fixed parameters
}
\examples{
power_calc()

}
