% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_invivoSyn_app}
\alias{run_invivoSyn_app}
\title{Run the invivoSyn Shiny application}
\usage{
run_invivoSyn_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
The value returned by \code{\link[shiny:runApp]{shiny::runApp()}}.
}
\description{
Run the invivoSyn Shiny application
}
