% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_colour_Publication}
\alias{scale_colour_Publication}
\title{Scale color for publication}
\usage{
scale_colour_Publication(...)
}
\arguments{
\item{...}{}
}
\value{
Scale color for publication
}
\description{
Scale color for publication
}
\examples{
scale_colour_Publication()
}
