% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_fill_Publication}
\alias{scale_fill_Publication}
\title{Scale fill for publication}
\usage{
scale_fill_Publication(...)
}
\arguments{
\item{...}{}
}
\value{
Scale fill for publication
}
\description{
Scale fill for publication
}
\examples{
scale_fill_Publication()
}
