% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{sim_power}
\alias{sim_power}
\title{Calculate power curve for specific paramter combination)}
\usage{
sim_power(...)
}
\arguments{
\item{...}{named parameters, same as power_calc}
}
\value{
a dataframe of powers for specific paramter combination
}
\description{
Calculate power curve for specific paramter combination)
}
\examples{
power_stats=sim_power(method=c('Bliss','HSA'),type=c('TGI','AUC','lmm'),n_sim=10,n=5:6)
}
