% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{TGI_synergy}
\alias{TGI_synergy}
\title{Synergy calculation based on TGI (TGI can either be defined by 1-delta(t)/delta(c) or 1-RTV(t)/RTV(c))}
\usage{
TGI_synergy(
  TGI_lst,
  method = "Bliss",
  ci = 0.95,
  ci_type = "perc",
  display = TRUE,
  save = TRUE,
  file = "TGI_synergy",
  width = 16,
  height = 8
)
}
\arguments{
\item{TGI_lst}{An list object returned by getTGI function}

\item{method}{method for synergy calculation, can be Bliss,HSA or RA}

\item{ci}{confidence intervaln for TGI based synergy score, can be 0.95,0.9,0.8, etc}

\item{ci_type}{type of bootstrap confidence interval, can be stud,perc or bca}

\item{display}{whether or not display figure}

\item{save}{save image}

\item{file}{if save is TRUE, the name of output file}

\item{width}{width (in inches) of the saved figure}

\item{height}{height (in inches) of the saved figure}
}
\value{
a data frame of synergy scores (CI and Synergy Score) and their
bootstrap confidence intervals, in the same format as \code{\link[=AUC_synergy]{AUC_synergy()}}
}
\description{
Generalized to N-drug combinations: handles any number of single-agent
arms via the \code{roles} list carried in \code{TGI_lst$roles}.
}
\examples{
TGI_lst=getTGI(LS_1034,17)
bliss_synergy_TGI=TGI_synergy(TGI_lst)
}
