% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Theme for publication}
\usage{
theme_Publication(base_size = 14, base_family = "helvetica")
}
\arguments{
\item{base_size}{text base size for ggplot}

\item{base_family}{text base family}
}
\value{
a theme for publication
}
\description{
Theme for publication
}
\examples{
theme_Publication()
}
