% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{bs_AUC_synergy}
\alias{bs_AUC_synergy}
\title{AUC based synergy calculation for bootstrap}
\usage{
bs_AUC_synergy(auc_mouse, t = 21, method = "Bliss", idx)
}
\arguments{
\item{auc_mouse}{auc dataframe for each mouse}

\item{t}{date to estimate survival (default 21 days)}

\item{method}{method for synergy calculation, can be Bliss,HSA or RA}

\item{idx}{bootstrap index}
}
\value{
combination index (CI) and synergy score
}
\description{
AUC based synergy calculation for bootstrap
}
