% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{bs_global_CI}
\alias{bs_global_CI}
\title{Bootstrap function to Calculate global CI based on TGI definition}
\usage{
bs_global_CI(tv_wide, method = "Bliss", idx)
}
\arguments{
\item{tv_wide}{tumor growth data in wide format (can be eiter deltaTV or RTV (mouse,Treatment,Group,Day...)), for bootstrap mouse within group}

\item{method}{method for synergy calculation, can be Bliss,HSA or RA}

\item{idx}{bootstrap index}
}
\value{
a dataframe of synergy scores and its bootstrap confidence interval
}
\description{
Bootstrap function to Calculate global CI based on TGI definition
}
