% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy.R
\name{bs_mAUCr}
\alias{bs_mAUCr}
\title{Calculate medianAUCRatio statistic for bootstrap}
\usage{
bs_mAUCr(auc_df, ref_group = "Group 1", idx)
}
\arguments{
\item{auc_df}{AUC data frame for each mouse}

\item{ref_group}{reference group}

\item{idx}{bootstrap index}
}
\value{
A vector of medianAUCRatio statistic for each group
}
\description{
Calculate medianAUCRatio statistic for bootstrap
}
