% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy.R
\name{bsTGI}
\alias{bsTGI}
\title{Calculate TGI statistic for bootstrap}
\usage{
bsTGI(tvs, tv_var = "DeltaTV", ref_group = "Group 1", idx)
}
\arguments{
\item{tvs}{tumor growth data of one specific day}

\item{tv_var}{variable for TGI calculation, can be either DeltaTV or RTV (RTV: definiton of CombPDX for TGI)}

\item{ref_group}{reference group,by default the first group (vehicle)}

\item{idx}{bootstrap index}
}
\value{
TGI statistics for each group
}
\description{
Calculate TGI statistic for bootstrap
}
