% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{CombPDX_CI}
\alias{CombPDX_CI}
\title{Calculate Day-specifc CI based on CombPDX's method}
\usage{
CombPDX_CI(tv, sel_day = NA, method = "Bliss", ci = 0.95)
}
\arguments{
\item{tv}{tumor growth data}

\item{sel_day}{the specific day for CI calcuation}

\item{method}{method for synergy calculation, can be Bliss,HSA or RA}

\item{ci}{confidence interval for CombPDX global CI, can be 0.95,0.9,0.8, etc;standard error based on delta method}
}
\value{
a vector of CI and it's confidence interval
}
\description{
Calculate Day-specifc CI based on CombPDX's method
}
