% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tv.R
\name{expand_tv}
\alias{expand_tv}
\title{Expand tumor volume data frame, add TV0, RTV and logTV}
\usage{
expand_tv(tv_long)
}
\arguments{
\item{tv_long}{tumor volume data frame, including Treatment, Mouse, Day and TV}
}
\value{
a tumor volume data frame, including TV,RTV and deltaTV
}
\description{
Expand tumor volume data frame, add TV0, RTV and logTV
}
\examples{
tv_long=expand_tv(tv_long)
}
