% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy.R
\name{get_mAUCr}
\alias{get_mAUCr}
\title{Calculate median AUC ratio using bootstrap}
\usage{
get_mAUCr(
  tv,
  sel_day = NA,
  ref_group = "Group 1",
  ci = 0.95,
  ci_type = "perc",
  nrep = 1000
)
}
\arguments{
\item{tv}{tumor growth data}

\item{sel_day}{the last day selected for tumor growth data, if not defined, use all data}

\item{ref_group}{reference group}

\item{ci}{confidence interval}

\item{ci_type}{type of bootstrap confidence interval, can be stud,perc or bca}

\item{nrep}{number of bootstrap replcates}
}
\value{
A list of boostrap mAUCr and mAUCr dataframe
}
\description{
Calculate median AUC ratio using bootstrap
}
\examples{
get_mAUCr(LS_1034,ci=0.9,ci_type='bca')
}
