% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy.R
\name{getTGI}
\alias{getTGI}
\title{Calculate TGI using stratified bootstrap,with confidence interval}
\usage{
getTGI(
  tv,
  sel_day,
  tv_var = "DeltaTV",
  ref_group = "Group 1",
  ci = 0.95,
  ci_type = "perc",
  n_rep = 1000
)
}
\arguments{
\item{tv}{object return from read_tv}

\item{sel_day}{timepoint for TGI calculation}

\item{tv_var}{variable for TGI calculation, can be either DeltaTV or RTV (RTV: definiton of CombPDX for TGI)}

\item{ref_group}{name of reference group}

\item{ci}{confidence interval, can be 0.95,0.9,0.8, etc}

\item{ci_type}{type of bootstrap confidence interval, can be stud,perc or bca}
}
\value{
a list of bootstrap result and TGI statistic (TGI and its confidence interval for each group)
}
\description{
Calculate TGI using stratified bootstrap,with confidence interval
}
\examples{
getTGI(LS_1034,17,ci=0.9,ci_type='bca')
}
