% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{lmm_synergy}
\alias{lmm_synergy}
\title{Calculate synergy based on linear mixed model}
\usage{
lmm_synergy(tv, sel_day = NA)
}
\arguments{
\item{tv}{tumor growth data}

\item{sel_day}{the last day selected for tumor growth data, if not defined, use all data}
}
\value{
a data frame, linear mixed model results
}
\description{
Calculate synergy based on linear mixed model
}
\examples{
lmm_synergy(LS_1034)
}
