% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tv.R
\name{read_tv}
\alias{read_tv}
\title{read_tv}
\usage{
read_tv(file)
}
\arguments{
\item{file}{the path of input file in a wide format, should have Treatment, Mouse and days in numeric number,
treatment should be in the order of vehicle,group 1, group 2, and combo}
}
\value{
a tumor volume data frame, including TV,RTV and deltaTV
}
\description{
read_tv
}
\examples{
library(SynergyInVivo)
read_tv(system.file("extdata","test.csv",package="SynergyInVivo"))
}
