% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{TGI_synergy}
\alias{TGI_synergy}
\title{Synergy calculation based on TGI (TGI can either be defined by 1-delta(t)/delta(c) or 1-RTV(t)/RTV(c))}
\usage{
TGI_synergy(
  TGI_lst,
  method = "Bliss",
  ci = 0.95,
  ci_type = "perc",
  display = TRUE,
  save = TRUE
)
}
\arguments{
\item{TGI_lst}{An list object returned by getTGI function}

\item{method}{method for synergy calculation, can be Bliss,HSA or RA}

\item{ci}{confidence intervaln for TGI based synergy score, can be 0.95,0.9,0.8, etc}

\item{ci_type}{type of bootstrap confidence interval, can be stud,perc or bca}

\item{display}{whether or not display figure}

\item{save}{save image}
}
\value{
Result of synergy calculation
}
\description{
Synergy calculation based on TGI (TGI can either be defined by 1-delta(t)/delta(c) or 1-RTV(t)/RTV(c))
}
\examples{
TGI_lst=getTGI(LS_1034,17)
bliss_synergy_TGI=TGI_synergy(TGI_lst)
}
